Add cascade flow calculations
This commit adds thbe necessary calculations to determine the product and
tails flows of the cascade. It includes a new constructor and the
`EnrichmentOutput` function has been adapted, as well.

In future commits, constraints on the available SWU and on the available
feed quantity need to be implemented. Moreover, the SWU calculations in
general need to be implemented.