Add assignment operator, remove ambiguous constructor
In addition to the above mentioned changes, a major typo has been fixed.
In EnrichmentCalculator::CalculateNStages, the `<` in the if clause
```c++
if (std::fmod(n_stages, 1.) < 1e-9)
```
has been corrected and now is a 'greater than', because the error should
only be thrown if the value is _not_ an integer (up to `1e-9`).