Change helper calculations from `alpha` to `gamma`
In previous versions, `multiisotopehelper::CalculateSeparationFactor`
expected to get `alpha`, the product-to-feed separation factor, as an
argument. This behaviour has been changed and from now on `gamma`, the
overall separation factor, is passed.

This change is made as the product-to-feed separation factor is nowhere
needed in the calculations and usually, the overall separation factor (being
the more "encompassing" factor) is indicated.