Move `EXPECT_TRUE_compmap` to `miso_helper`
The function `EXPECT_TRUE_compmap` has been moved from
`enrichment_calculator` to `miso_helper` as it is a helper function to be
used during unit testing. Moving it to said file is logical as it literally
is a helper function. The already existing code has been adapted.

Moreover, cosmetics have done to the helper file (making the horizontal
delimiters, `// - - - - `, consistent with other files).