Add global constants, fix `EXPECT_TRUE_compmap`
Global constants, such as the maximum number of iterations or
epsilon/threshold values used for comparing doubles, have been added to the
helper file. The list might be extended in the future. The adaptations
needed to all of the other files have been made.

Moreover, `EXPECT_TRUE_compmap` has been fixed. The changes implemented in
eefd813cbad0cb64cb3fecaa15b2d9365b3dac22 did not compile proper and yielded
a linking error. To prevent this, the function does not use the actual
gtest `EXPECT_TRUE` function but only performs the comparison and then
returns a `bool` value which should be passed to gtest's `EXPECT_TRUE`.