Change n_stages to int in enrichment_calculator, add pprint
Previously, `n_enriching` and `n_stripping` have been declared as doubles
because it was handy while programming. However, the `CalculateNStages_`
function has been redefined such that it also works with the number of
stages being stored as ints. This is conceptually clearer than using
integer numbers but storing them as doubles.

In addition to this change, a prettyprint function has been added. It will
not be used in the CYCLUS simulation but it can come in handy during
debugging or for the calculation of enrichments.