Fix enrichment calculator unit tests
All but one unit tests of `EnrichmentCalculator` are now running. Naming
has been made consistent, the required feed quantity for an enrichment is
now checked as well and for the flow unit test `EXPECT_DOUBLE_EQ` has been
replaced by `EXPECT_NEAR` to account for rounding errors.

The Swu unit test is failing which is expected as the value function has
not yet been implemented.