Check requested assay in MIsoEnrich::ValidReq
Previously, `MIsoEnrich::ValidReq_` checked if the requested enrichment was
valid solely by considering if the requested assay was larger than the
tails assay and if there was U238 present. Outside the function, it was
then checked if the requested assay did not exceed the facility's maximum
enrichment level.

This second check has been moved and is now performed by `ValidReq_`, as
well, since a non-feasible enrichment request also is an invalid request.

Also, the `MIsoEnrichTest` class (featuring the MIsoEnrich unit tests) has
been added as friend class to allow the testing of private functions.