Add `AddFeedMat_` to MIsoEnrich class
Previously, the facility would check if the feed composition to be added
did already exist or not. If not, it would try to push the material to a
non-existing `cyclus::ResBuf` which would of course lead to an error. This
has been fixed by creating the `AddFeedMat_` function.

It should be noted that for this purpose, the `MIsoHelper::ResBufIdx`
function has been adapted, as well.