Fix EnrichmentCalculator, add constructor and gamma235
This commit contains multiple changes:
  1. fix usage from within `MIsoEnrich`
  2. add constructors
  3. adapt `SetInput`
  4. add `CalculateGammaAlphaStar_`
  5. fix naming in unit tests for consistency

Changes 2. and 3. are part of change 1., change 4. has been added for
clarity and modularity, change 5. is independent of all of the above.

When used from within MIsoEnrich, the calculator would yield strange
results including negative stage separation factors and `nan` values. The
root of this problem is that `gamma_235` has not been properly initialised
giving wrong separation factors and therefore also wrong concentrations and
flows.

To fix this problem, the constructor of MIsoEnrich has been changed (not
part of this commit) such that it calls a constructor of
EnrichmentCalculator (new as well, part of this commit) which sets
`gamma_235` and calculates all the separation factors. Moreover, each
constructor of `EnrichmentCalculator` now correctly initialises and fills
the `isotopes` vector.