Use Agent as parent instead of Facility, fix minor bugs
From this commit on, the parent does not have to be a cyclus::Facility but
it has to be cyclus::Agent (which includes Facility and also Institution
and Region). Additionally, it prevents typecasting in the unit tests (the
unit tests are not yet committed).

Moreover, minor bugs and typos have been edited.