Add FlexibleInput to MIsoenrich, fix FlexibleInput.UpdateValue()
While testing it has been found that the `parent_` pointer in FlexibleInput
does not work reliably, i.e., it does not point to its parent. This might
be due to the pointer being initialised and assigned and then it is deleted
again during start-up of Cyclus (as indicated when using cyclus in verbose
mode, see the following piece of code)
```
DEBUG3(memory):              Deleting agent 'Foo' ID=23
DEBUG3(memory):              Agent ID=33, ptr=0x7f924a18ac88 created.
```
To fix this, the parent or its pointer is not saved in the file, it is
always passed. Its enter time is stored.

The last used time and value entry are referred using an iterator over the
vector.

While incorporating FlexibleInput into MIsoEnrich, it has been found that
the `this` pointer is not correctly passed to FlexibleInput. For unknown
reasons, the address was always off by 8 bytes. To fix this, the address of
the pointer is copied using `std::memcpy` and then passed. Relevant other
code (such as pragmas for configuring FlexibleInput) has been incorporated,
as well.