Adapt MIsoEnrich unit tests to incorporate FlexibleInput
As `FlexibleInput` uses the duration of the simulation or the lifetime of
the agent, changes had to be made to the simulation context to prevent
pointer errors. `MockSim` is used to create a working and well-defined
simulation context which is passed to `MIsoEnrich`. Moreover, the input
strings to mocksims have been adapted to the new input variable.