Add downblending method to `EnrichmentCalculator`
This method allows to reach precise enrichment levels (up to around 0.05%
precision): The enrichment is done in discrete stages, meaning that only
discrete increments in a product's enrichment level can be done. The down-
blend method uses uranium from the feed, mixes as much feed U as needed
with the product to obtain the desired enrichment level.

It should be noted that all of this is done in the enrichment facility in
Cyclus and it will not be registered separately in a different log.