Fix bugs in enrichment_calculator, update unit tests
In `Downblend_`, the calculation of the downblend feed quantity has been
updated. Previously, in the case of feed constraints, the formula used was
a rather crude approximation that worked _most_ of the times (i.e., it did
fail sometimes). This has been updated and now works correctly. The unit
tests have been updated, as well.

A small but major bug was fixed in `SetInput` where the `use_downblending`
variable was not properly updated.

Also, the calculation of the sums from von Halle's paper, Eqs. (47) and
(50), has been outsourced to a separate function to maintain a clear code.

Finally, the unit tests of MIsoEnrich have been updated, see
`miso_enrich_tests.cc`.