Remove unnecessary helper function
Two functions existed in `MIsoHelper` that would calculate the 'atomic'
and 'mass' fractions in a `cyclus::CompMap` object. However, this is
non-sensical as only `cyclus::Composition` objects can be subdivided in a
'mass' and in an 'atom' composition part. The functions and all of the
function calls have been adapted.