Fix bug in EnrichmentCalculator::Downblend_()
Minor discrepencies in the use of Downblend_() were noticed.
The error has been fixed by adding the blending feed to the product and
feed quantities _after_ the downblended product composition has been
calculated.

This can be understood when looking at the calculation of the product
composition. It is the sum of product and feed concentration weighted with
their mass. Hence the blending feed mass needs to be added _after_ the
downblending.