WIP: Add gaseous diffusion as enrichment method
The separation factors used in enrichment by gaseous diffusion differ from
the ones used in enrichment by gas centrifugation. This difference is
accounted for in this commit.

WIP: currently the unit tests pass but (one) test simulation did not start
properly. This needs further investigation, first to determine if the
failure stems from the new implementation or if it results of the input
file (which it shouldn't, though, this input file has already been used in
different examples).