WIP Add untested `predict_posterior` Python script
`predict_posterior.py` is now in a preliminary final version (up to
potential changes of paths or variable names) but it has not been tested
yet.

`kernel.py` contains lots of kernel types and this file will be cleaned
up in the future, all unneeded kernel types will be removed. However,
this can only be done after the kernels have been trained and after it
is known which kernel type is 'the chosen one'.